+++ packages/generate-project/src/ejsRenderDir.ts
@@ -1,10 +1,14 @@
 import * as path from 'path';
-import * as glob from 'glob';
 import * as ejs from 'ejs';
 import * as fse from 'fs-extra';
+import type { Data as ejsData } from 'ejs';
+import formatFileContent from './formatFileContent';
+import type { Options } from 'prettier';
 
-export default async function (dir: string, options: any): Promise<void> {
-  return new Promise((resolve, reject) => {
+import glob = require('glob');
+
+export default async function ejsRenderDir(dir: string, data: ejsData): Promise<void> {
+  const files: string[] = await new Promise((resolve, reject) => {
     glob(
       '**/*.ejs',
       {
@@ -13,38 +17,43 @@ export default async function (dir: string, options: any): Promise<void> {
         dot: true,
         ignore: ['node_modules/**'],
       },
-      (err, files) => {
-        if (err) {
-          return reject(err);
+      (error, matches) => {
+        if (error) {
+          reject(error);
         }
-
-        Promise.all(
-          files.map((file) => {
-            const filepath = path.join(dir, file);
-            return renderFile(filepath, options);
-          }),
-        )
-          .then(() => {
-            resolve();
-          })
-          .catch((error) => {
-            reject(error);
-          });
+        resolve(matches);
       },
     );
   });
+
+  await Promise.all(
+    files.map((file) => {
+      const filepath = path.join(dir, file);
+      return renderAndFormatFile(filepath, data);
+    }),
+  );
 }
 
-function renderFile(filepath: string, options: any): Promise<string> {
-  return new Promise((resolve, reject) => {
-    ejs.renderFile(filepath, options, (err, result) => {
-      if (err) {
-        return reject(err);
-      }
+const parserMaps: Array<[RegExp, string]> = [
+  [/\.json$/, 'json'],
+  [/\.(ts|tsx|mts|cts)$/, 'babel-ts'],
+  [/\.(js|jsx|mjs|cjs)$/, 'babel'],
+  [/\.(yml|yaml)$/, 'yaml'],
+];
 
-      fse.removeSync(filepath);
-      fse.writeFileSync(filepath.replace(/\.ejs$/, ''), result);
-      resolve();
-    });
-  });
+async function renderAndFormatFile(filepath: string, data: ejsData): Promise<void> {
+  const fileContent = await ejs.renderFile(filepath, data);
+  const realFilepath = filepath.replace(/\.ejs$/, '');
+
+  const formatOptions: Options = {};
+
+  for (const [regexp, parser] of parserMaps) {
+    if (regexp.test(realFilepath)) {
+      formatOptions.parser = parser;
+      continue;
+    }
+  }
+
+  await fse.writeFile(realFilepath, formatFileContent(fileContent, formatOptions));
+  await fse.remove(filepath);
 }